requirements.txt
openai==1.59.8
python-dotenv==1.0.1
streamlit==1.44.0
av==14.2.0
faster-whisper==1.1.1
pyaudio==0.2.14
pygame==2.6.1
scipy==1.15.2
numpy==2.2.4
edge-tts==7.0.0
ctranslate2==4.5.0